Updated databases, added routes and started UI and functionality

+++ database/migrations/2020_06_04_161024_create_transactions_table.php
@@ -20,7 +20,7 @@ class CreateTransactionsTable extends Migration
             $table->foreignId('added_by_user_id')->constrained('users');
             $table->foreignId('income_source_id')->nullable()->constrained();
             $table->boolean('is_income');
-            $table->integer('amount');
+            $table->decimal('amount', 12, 2);
             $table->string('description');
             $table->string('image_uuid')->nullable();
             $table->timestamps();
